chore: Refactoring project structure and tab switching
