+++ client/src/lib/containers/intercept.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
-    import H1 from "$lib/H1.svelte";
+    import H1 from "$lib/components/H1.svelte";
 
-    const pageTitle: string = "Options";
+    const pageTitle: string = "Intercept";
 </script>
 
 <svelte:head>
