+++ client/src/routes/__layout.svelte
@@ -1,23 +1,12 @@
 <script lang="ts">
     import Header from "$lib/Header.svelte";
-    import TabGroup from "$lib/TabGroup.svelte";
 
     let title: string = "My Cool Project";
-
-    let tabs = [
-        {title: "Intercept", active: true},
-        {title: "History", active: false},
-        {title: "Options", active: false},
-    ];
 </script>
 
 <Header title={title} />
 <div class="layout">
-    <TabGroup {tabs} />
-
-    <div class="view">
-        <slot />
-    </div>
+    <slot />
 </div>
 
 <style>
@@ -26,7 +15,4 @@
         padding: 24px 0px;
     }
 
-    .view {
-        padding: 24px 0 0 0;
-    }
 </style>
